Put Google Analytics stuff so the functional route renders both on server/client
